move auth helpers to pkg/server, remove pkg/memory and pkg/kagentapi
